requirements.txt
flask
redis
apscheduler
numpy
bayesian-testing
python-dotenv